webhook_requirements.txt
python-fasthtml
requests==2.31.0
google-generativeai==0.8.3
Pillow==10.0.1
